+++ src/main/java/com/shinobicoders/teamcodeapi/controller/UserController.java
@@ -3,13 +3,15 @@ package com.shinobicoders.teamcodeapi.controller;
 import com.shinobicoders.teamcodeapi.model.User;
 import com.shinobicoders.teamcodeapi.service.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.security.core.context.SecurityContextHolder;
 
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/api/v1/user")
@@ -30,14 +32,30 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User userDetails) {
+    public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User userDetails) throws ChangeSetPersister.NotFoundException {
+        // Get principal
+        UserDetails principal = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User user = userService.getUserByEmail(principal.getUsername());
+
+        // Check if user is deleting their own account
+        if (!Objects.equals(user.getId(), id)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         User updatedUser = userService.updateUser(id, userDetails);
         return new ResponseEntity<>(updatedUser, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable Long id) {
-        // todo: check auth
+    public ResponseEntity<?> deleteUser(@PathVariable Long id) throws ChangeSetPersister.NotFoundException {
+        // Get principal
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User user = userService.getUserByEmail(userDetails.getUsername());
+
+        // Check if user is deleting their own account
+        if (!Objects.equals(user.getId(), id)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
 
         userService.deleteUser(id);
         return new ResponseEntity<>(HttpStatus.OK);
